+++ src/components/registerForm/cardNumber/CardNumber.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef, useState, useContext, useRef } from 'react';
 
 import styled, { css } from 'styled-components';
 
-import { InputValuesContext } from '../InputValueContext';
+import { InputValuesContext } from '../../../contexts/InputValueContext';
 import useAutoFocus from '../../../hooks/useAutoFocus';
 import FormLabel from '../../@common/FormLabel';
 import Input from '../../@common/Input';
@@ -45,13 +45,17 @@ export const CardNumber = forwardRef<HTMLDivElement, Props>(({}, ref) => {
     const inputIndex = event.currentTarget.dataset['index'];
     if (!order) return;
 
+    // 숫자가 아니면 입력받지 않는다.
     if (!ONLY_NUMBER_REGEXP.test(inputValue)) return;
 
     try {
+      // validation1. 4글자인지 확인하기
+      // FIXME blur 되었을 때 (border처리 및 메세지 출력 및 focus 유지)
       if (inputValue.length !== 4) {
         throw new Error(`4글자를 입력해 주세요`);
       }
 
+      // validation 통과하면 에러메세지 없애기
       setCardError({
         isError: false,
         message: '',
