+++ src/components/registerForm/expireDate/ExpireDate.tsx
@@ -2,7 +2,7 @@ import React, { useState, useContext } from 'react';
 
 import styled, { css } from 'styled-components';
 
-import { InputValuesContext } from '../InputValueContext';
+import { InputValuesContext } from '../../../contexts/InputValueContext';
 import { MMYY_REGEXP, ONLY_NUMBER_REGEXP } from '../../../utils/regexp';
 import FormLabel from '../../@common/FormLabel';
 import Input from '../../@common/Input';
